minor improvements - hide ai suggestions  when not in drafting
